+++ src/lib/data/planned-releases.ts
@@ -5,6 +5,47 @@ import { Session } from "next-auth";
 import { Track } from "@/types/tracks";
 import { getSongInfoMultiple } from "./spotify/track";
 
+export async function changePlannedReleaseDate(
+  plannedReleaseId: string,
+  date: Date,
+  session: Session
+) {
+  if (!session?.user || !session.user.id) {
+    throw new Error("User not found");
+  }
+
+  const plannedRelease = await prisma.plannedRelease.findFirst({
+    where: {
+      id: plannedReleaseId,
+      managedPlaylist: {
+        userId: session.user.id,
+      },
+    },
+  });
+
+  if (!plannedRelease) {
+    throw new Error("Planned release not found");
+  }
+
+  if (
+    plannedRelease.scheduledAt &&
+    plannedRelease.scheduledAt.toISOString().split("T")[0] <=
+      new Date().toISOString().split("T")[0]
+  ) {
+    throw new Error("Planned release has once already been activated");
+  }
+  const updatedPlannedRelease = await prisma.plannedRelease.update({
+    where: {
+      id: plannedReleaseId,
+    },
+    data: {
+      scheduledAt: date,
+    },
+  });
+
+  return updatedPlannedRelease;
+}
+
 /**
  * create a planned release and redirects to the planning page for further editing
  * @param managedPlaylistId string
@@ -67,6 +108,10 @@ export async function deletePlannedRelease(
     throw new Error("Planned release not found");
   }
 
+  if (plannedRelease.isActivated) {
+    throw new Error("Planned release is activated and cannot be deleted");
+  }
+
   await prisma.plannedRelease.delete({
     where: {
       id: plannedReleaseId,
@@ -96,6 +141,17 @@ export default async function getPlannedReleases(
         spotifyPlaylistId: playlistId,
       },
     },
+    orderBy: [
+      {
+        isActivated: "desc",
+      },
+      {
+        scheduledAt: "desc",
+      },
+      {
+        createdAt: "desc",
+      },
+    ],
     include: {
       plannedSongs: {
         take: 5,
